Add node selector and HTTP request node support
Introduces a NodeSelector component for adding nodes via a UI, updates the schema and migrations to support MANUAL_TRIGGER and HTTP_REQUEST node types, and implements the HTTP Request node component. Replaces the previous AddNodeButton and editor-node-button with a new version integrated with the NodeSelector. Updates node-components config to register the new node type.